Fixed volume renderer Output # of voxels instead of logical size.

+++ trrojancore/src/brudervn_xfer_func.cpp
@@ -17,14 +17,14 @@
 std::vector<std::uint8_t> trrojan::load_brudervn_xfer_func(
         const std::string& path) {
     std::vector<std::uint8_t> retval;
-    std::uint8_t value;
+    std::uint32_t value;
 
     std::ifstream file(path, std::ios::in);
     if (file.is_open()) {
         retval.reserve(256);
 
         while (file >> value) {
-            retval.push_back(value);
+            retval.push_back(static_cast<std::uint8_t>(value));
         }
 
     } else {
